requirements.txt
Django==3.0
Scrapy==2.5.0
requests==2.25.1
beautifulsoup4==4.9.3
Pillow==8.2.0
python-slugify==5.0.2
djangorestframework==3.12.4
